# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAR Fleet Service",
    "summary": "Ground fleet service for civilian search-and-rescue drone operations.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/aircraft": {
      "get": {
        "tags": [
          "aircraft"
        ],
        "summary": "List Aircraft",
        "description": "List registered aircraft in registration order.",
        "operationId": "list_aircraft_api_v1_aircraft_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum items to return."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous page's next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous page's next_cursor."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftPage"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "fleet.view"
      },
      "post": {
        "tags": [
          "aircraft"
        ],
        "summary": "Create Aircraft",
        "description": "Register an aircraft.",
        "operationId": "create_aircraft_api_v1_aircraft_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AircraftCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftOut"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "fleet.manage"
      }
    },
    "/api/v1/aircraft/{aircraft_id}": {
      "get": {
        "tags": [
          "aircraft"
        ],
        "summary": "Get Aircraft",
        "description": "One aircraft.",
        "operationId": "get_aircraft_api_v1_aircraft__aircraft_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "aircraft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aircraft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftOut"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "fleet.view"
      },
      "patch": {
        "tags": [
          "aircraft"
        ],
        "summary": "Update Aircraft",
        "description": "Change an aircraft's registration.",
        "operationId": "update_aircraft_api_v1_aircraft__aircraft_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "aircraft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aircraft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AircraftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AircraftOut"
                }
              }
            }
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "fleet.manage"
      },
      "delete": {
        "tags": [
          "aircraft"
        ],
        "summary": "Delete Aircraft",
        "description": "Remove an aircraft from the registry; refused while any task refers to it.",
        "operationId": "delete_aircraft_api_v1_aircraft__aircraft_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "aircraft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aircraft Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "The request body could not be parsed (not valid JSON or not UTF-8).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The session's role lacks the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input: malformed, out of range, or semantically invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-permission": "fleet.manage"
      }
    },
...
  },
  "components": {
    "schemas": {
      "AircraftCreate": {
        "properties": {
          "callsign": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]{0,31}$",
            "title": "Callsign",
            "description": "Radio callsign, stored upper-case; unique ignoring case."
          },
          "airframe": {
            "$ref": "#/components/schemas/Airframe"
          },
          "mavlink_system_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 254.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mavlink System Id"
          },
          "mavlink_connection": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "pattern": "^(udp|udpin|udpout|tcp|tcpin|tcpout|serial)://\\S+$",
                "description": "MAVSDK connection string, e.g. udp://:14541 (ADR 0010)."
              },
              {
                "type": "null"
              }
            ],
            "title": "Mavlink Connection"
          },
          "swarm_drone_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "description": "drone_id in the onboard swarm protocol."
              },
              {
                "type": "null"
              }
            ],
            "title": "Swarm Drone Id"
          },
          "cruise_speed_mps": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 60.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cruise Speed Mps"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "callsign",
          "airframe"
        ],
        "title": "AircraftCreate",
        "description": "A new aircraft. Links may be added later; a task needs the matching link (ADR 0003)."
      },
      "AircraftOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "callsign": {
            "type": "string",
            "title": "Callsign"
          },
          "airframe": {
            "$ref": "#/components/schemas/Airframe"
          },
          "mavlink_system_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mavlink System Id"
          },
          "mavlink_connection": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mavlink Connection"
          },
          "swarm_drone_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swarm Drone Id"
          },
          "cruise_speed_mps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cruise Speed Mps"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "group_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Group Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "callsign",
          "airframe",
          "mavlink_system_id",
          "mavlink_connection",
          "swarm_drone_id",
          "cruise_speed_mps",
          "notes",
          "group_ids",
          "created_at",
          "updated_at"
        ],
        "title": "AircraftOut",
        "description": "A registered aircraft."
      },
      "AircraftPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AircraftOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Null when this is the last page."
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AircraftPage",
        "description": "A page of aircraft."
      },
      "AircraftUpdate": {
        "properties": {
          "callsign": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]{0,31}$",
            "title": "Callsign",
            "description": "Radio callsign, stored upper-case; unique ignoring case."
          },
          "airframe": {
            "$ref": "#/components/schemas/Airframe"
          },
          "mavlink_system_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 254.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mavlink System Id"
          },
          "mavlink_connection": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "pattern": "^(udp|udpin|udpout|tcp|tcpin|tcpout|serial)://\\S+$",
                "description": "MAVSDK connection string, e.g. udp://:14541 (ADR 0010)."
              },
              {
                "type": "null"
              }
            ],
            "title": "Mavlink Connection"
          },
          "swarm_drone_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "description": "drone_id in the onboard swarm protocol."
              },
              {
                "type": "null"
              }
            ],
            "title": "Swarm Drone Id"
          },
          "cruise_speed_mps": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 60.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cruise Speed Mps"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AircraftUpdate",
        "description": "Fields to change; ``null`` clears a nullable field."
      },
      "Airframe": {
        "type": "string",
        "enum": [
          "fixed_wing",
          "multirotor_hexa",
          "multirotor_quad"
        ],
        "title": "Airframe",
        "description": "Aircraft type; drives planning constraints (turn radius, hover ability)."
      },
...
      "FieldError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc",
            "description": "Where the value is: body, query, path, header."
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FieldError",
        "description": "One invalid input value."
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank",
            "examples": [
              "urn:sar-gcs:problem:conflict"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Present on validation errors: each invalid value."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details. Extension members depend on ``type``."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Session token returned by POST /api/v1/auth/login.",
        "scheme": "bearer"
      }
    }
  },
...
}